+++ src/components/Main/css/LoadingCard.module.css
